% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle_bc_R.R
\name{BC.APMLE.R}
\alias{BC.APMLE.R}
\title{Compute The Bias Correction After The L1 Penalized Maximum Likelihood Estimator}
\usage{
BC.APMLE.R(x, y, lambda0, beta.pmle, sigma.pmle)
}
\arguments{
\item{x}{the design matrix}

\item{y}{the response vector}

\item{lambda0}{the pre-fixed penalty level}

\item{beta.pmle}{penalized maximum likelihood estimator of beta}

\item{sigma.pmle}{penalized maximum likelihood estimator of sigma}
}
\value{
beta.bc and sigma.bc
}
\description{
More details refer to scaled sparse linear regression
}
\examples{
\dontrun{
    r1.pmle<-PMLE.Iter.R(d1[,-c(1,2)],d1[,1001],lambda0=d1[1,1002])
    BC.APMLE.R(
      d1[,-c(1,2)],
      d1[,1001],
      lambda0=d1[1,1002],
      r1.pmle$beta.pmle,
      r1.pmle$sigma.pmle
    )
}
}
