% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ratio_cpp}
\alias{compute_ratio_cpp}
\title{Compute sigma.hat/sigma.star for evaluation}
\usage{
compute_ratio_cpp(x, y, beta_star, sigma_hat)
}
\arguments{
\item{x}{the design matrix}

\item{y}{the response vector}

\item{beta_star}{the true coefficient vector of beta}

\item{sigma_hat}{the estimator of sigma}
}
\value{
sigma_hat/sigma_star
}
\description{
More details refer to scaled sparse linear regression
}
\examples{
\dontrun{
   compute_ratio_cpp(d1[,1:1000],as.vector(d1[,1001]),rep(0,1000),1)
}
}
