% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle_bc_R.R
\docType{data}
\name{d1}
\alias{d1}
\title{A illustration dataset with r0=0}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
d1
}
\description{
A dataset with r0=0 used to illustrate the performance of pmle and bc of versions R and Rcpp
}
\examples{
\dontrun{
    data(d1)
    attach(d1)
}
}
\keyword{datasets}
