% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle_bc_R.R
\docType{data}
\name{d2}
\alias{d2}
\title{A illustration dataset with r0=0.5}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
d2
}
\description{
A dataset with r0=0.5 used to illustrate the performance of pmle and bc of versions R and Rcpp
}
\examples{
\dontrun{
    data(d2)
    attach(d2)
}
}
\keyword{datasets}
