% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmle_bc_R.R
\name{PMLE.Iter.R}
\alias{PMLE.Iter.R}
\title{Compute The L1 Penalized Maximum Likelihood Estimator}
\usage{
PMLE.Iter.R(x, y, lambda0, max.it = 5000)
}
\arguments{
\item{x}{the design matrix}

\item{y}{the response vector}

\item{lambda0}{the pre-fixed penalty level}

\item{max.it}{the maximum number of iterations}
}
\value{
beta.pmle and sigma.pmle
}
\description{
More details refer to scaled sparse linear regression
}
\examples{
\dontrun{
    PMLE.Iter.R(d1[,-c(1,2)],d1[,1001],lambda0=d1[1,1002])
    PMLE.Iter.R(d2[,-c(1,2)],d2[,1001],lambda0=d2[1,1002])
}
}
