\name{SA23204175-package}
\alias{SA23204175-package}
\docType{package}
\title{
  Final Project Of Statistical Computing
}
\description{
  Final project includes two functions of my research project
and all my homework answers.
}
\author{
Wang Xiaoli
Maintainer: Wang Xiaoli <wxl1026@mail.ustc.edu.cn>
}
\keyword{ package }
